+++ release-notes.md
@@ -4,6 +4,7 @@ Pending changes in the mainline
 Fixes:
   - [Issue #51](https://github.com/orthanc-server/orthanc-explorer-2/issues/51) Stop calling /changes once all studies are displayed.
   - [Issue #52](https://github.com/orthanc-server/orthanc-explorer-2/issues/52) Removed duplicate calls to /studies/../series and /series/../instances.
+  - [Issue #53](https://github.com/orthanc-server/orthanc-explorer-2/issues/53) Changing selection in Modalities in Study works immediately even if `"StudyListSearchMode": "search-button"`.
   - [Issue #54](https://github.com/orthanc-server/orthanc-explorer-2/issues/54) OHIF disappeared after #52 fix
   - [Issue #55](https://github.com/orthanc-server/orthanc-explorer-2/issues/55) DateFormat not respected + introduced DatePicker in the Modification Modal.
   - Removed duplicate calls to /tools/find in Modification modal